Move test vector test to integration test folder
